requirements.txt
# Web Scraping
requests>=2.31.0
beautifulsoup4>=4.12.0
selenium>=4.15.0
playwright>=1.40.0
lxml>=4.9.3

# LLM Integration
openai>=1.6.0
anthropic>=0.8.0
langchain>=0.1.0
langchain-openai>=0.0.2
tiktoken>=0.5.2

# Data Processing
pandas>=2.1.0
numpy>=1.24.0
pyyaml>=6.0.1
python-dotenv>=1.0.0

# Analysis & NLP
nltk>=3.8.1
spacy>=3.7.0
textblob>=0.17.1

# Scoring & Metrics
scikit-learn>=1.3.0
scipy>=1.11.0

# Visualization
matplotlib>=3.8.0
seaborn>=0.13.0
plotly>=5.18.0

# PDF Generation
reportlab>=4.0.0
fpdf2>=2.7.0
weasyprint>=60.0

# Data Export
openpyxl>=3.1.2
xlsxwriter>=3.1.9

# CLI
argparse>=1.4.0
colorama>=0.4.6
tqdm>=4.66.0

# Logging
loguru>=0.7.2

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mock>=3.12.0

# Code Quality
black>=23.12.0
flake8>=6.1.0
mypy>=1.7.0
